+++ src/main/kotlin/com/freewill/common/exception/BadRequestException.kt
@@ -1,4 +1,4 @@
-package com.freewill.global.common.exception
+package com.freewill.common.exception
 
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.ResponseStatus
