+++ src/main/kotlin/com/freewill/config/WebMvcConfig.kt
@@ -1,6 +1,6 @@
 package com.freewill.config
 
-import com.freewill.global.common.resolver.AuthorizedUserResolver
+import com.freewill.resolver.AuthorizedUserResolver
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.method.support.HandlerMethodArgumentResolver
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
