+++ src/main/kotlin/com/freewill/dto/param/UserRegisterParam.kt
@@ -1,6 +1,6 @@
-package com.freewill.domain.user.dto.param
+package com.freewill.dto.param
 
-import com.freewill.domain.user.dto.request.SignUpRequest
+import com.freewill.dto.request.SignUpRequest
 import org.springframework.web.multipart.MultipartFile
 
 data class UserRegisterParam(
