+++ src/main/kotlin/com/freewill/repository/CafeRepository.kt
@@ -1,6 +1,6 @@
-package com.freewill.domain.cafe.repository
+package com.freewill.repository
 
-import com.freewill.domain.cafe.entity.Cafe
+import com.freewill.entity.Cafe
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
