+++ src/main/kotlin/com/freewill/repository/GuestbookRepository.kt
@@ -1,6 +1,6 @@
-package com.freewill.domain.guestbook.repository
+package com.freewill.repository
 
-import com.freewill.domain.guestbook.entity.Guestbook
+import com.freewill.entity.Guestbook
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
