+++ src/main/kotlin/com/freewill/repository/ProfileRepository.kt
@@ -1,6 +1,6 @@
-package com.freewill.domain.profile.repository
+package com.freewill.repository
 
-import com.freewill.domain.profile.entity.Profile
+import com.freewill.entity.Profile
 import com.freewill.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
