+++ src/main/kotlin/com/freewill/repository/QuestionRepository.kt
@@ -1,6 +1,6 @@
-package com.freewill.domain.question.repository
+package com.freewill.repository
 
-import com.freewill.domain.question.entity.Question
+import com.freewill.entity.Question
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
