+++ src/main/kotlin/com/freewill/repository/RegionRepository.kt
@@ -1,6 +1,6 @@
-package com.freewill.domain.region.repository
+package com.freewill.repository
 
-import com.freewill.domain.region.entity.Region
+import com.freewill.entity.Region
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
