+++ src/main/kotlin/com/freewill/s3/S3Uploader.kt
@@ -1,11 +1,11 @@
-package com.freewill.global.s3
+package com.freewill.s3
 
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.AmazonS3Exception
 import com.amazonaws.services.s3.model.CannedAccessControlList
 import com.amazonaws.services.s3.model.ObjectMetadata
 import com.amazonaws.services.s3.model.PutObjectRequest
-import com.freewill.global.s3.dto.S3UploadRequest
+import com.freewill.s3.dto.S3UploadRequest
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import org.springframework.web.multipart.MultipartFile
