+++ src/main/kotlin/com/freewill/security/jwt/util/JwtValidator.kt
@@ -2,8 +2,8 @@ package com.freewill.security.jwt.util
 
 import com.fasterxml.jackson.core.JsonProcessingException
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.freewill.domain.user.entity.User
-import com.freewill.domain.user.service.UserService
+import com.freewill.entity.User
+import com.freewill.service.UserService
 import com.freewill.security.oauth.util.PrincipalUser
 import com.freewill.security.oauth.util.PrincipalUserConverter
 import io.jsonwebtoken.Claims
