+++ src/main/kotlin/com/freewill/security/oauth/util/PrincipalUserConverter.kt
@@ -1,6 +1,6 @@
 package com.freewill.security.oauth.util
 
-import com.freewill.domain.user.entity.User
+import com.freewill.entity.User
 import org.springframework.stereotype.Component
 
 @Component
