+++ src/test/kotlin/com/freewill/domain/controller/AuthControllerTest.kt
@@ -3,14 +3,14 @@ package com.freewill.domain.controller
 import com.freewill.docs.ApiDocumentUtils.getDocumentRequest
 import com.freewill.docs.ApiDocumentUtils.getDocumentResponse
 import com.freewill.docs.RestDocsTest
-import com.freewill.domain.user.controller.AuthController
-import com.freewill.domain.user.dto.param.OAuth2Param
-import com.freewill.domain.user.dto.param.UserRegisterParam
-import com.freewill.domain.user.dto.request.SignInRequest
-import com.freewill.domain.user.dto.request.SignUpRequest
-import com.freewill.domain.user.entity.User
-import com.freewill.domain.user.entity.enums.Provider
-import com.freewill.domain.user.service.AuthService
+import com.freewill.controller.AuthController
+import com.freewill.dto.param.OAuth2Param
+import com.freewill.dto.param.UserRegisterParam
+import com.freewill.dto.request.SignInRequest
+import com.freewill.dto.request.SignUpRequest
+import com.freewill.entity.User
+import com.freewill.enums.Provider
+import com.freewill.service.AuthService
 import com.freewill.security.jwt.dto.JwtToken
 import com.freewill.security.jwt.util.JwtValidator
 import org.junit.jupiter.api.BeforeEach
