+++ src/test/kotlin/com/freewill/domain/controller/ProfileControllerTest.kt
@@ -3,13 +3,13 @@ package com.freewill.domain.controller
 import com.freewill.docs.ApiDocumentUtils.getDocumentRequest
 import com.freewill.docs.ApiDocumentUtils.getDocumentResponse
 import com.freewill.docs.RestDocsTest
-import com.freewill.domain.profile.controller.ProfileController
-import com.freewill.domain.profile.dto.request.ProfileUpdateRequest
-import com.freewill.domain.profile.entity.Profile
-import com.freewill.domain.profile.service.ProfileService
-import com.freewill.domain.user.dto.param.OAuth2Param
-import com.freewill.domain.user.entity.User
-import com.freewill.domain.user.entity.enums.Provider
+import com.freewill.controller.ProfileController
+import com.freewill.dto.request.ProfileUpdateRequest
+import com.freewill.entity.Profile
+import com.freewill.service.ProfileService
+import com.freewill.dto.param.OAuth2Param
+import com.freewill.entity.User
+import com.freewill.enums.Provider
 import com.freewill.security.jwt.util.JwtValidator
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
